+++ src/main/java/ru/skillbox/socialnetwork/services/exceptions/PersonNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.skillbox.socialnetwork.services.exceptions;
 
 public class PersonNotFoundException extends RuntimeException {
-    public PersonNotFoundException(int id) {
+    public PersonNotFoundException(long id) {
         super("invalid profile ID: " + id);
     }
 }
